feat: 默认启用 GPU FP16 检索并优化 DINOv3 注意力 Pre-scale 计算
  - 图像特征检索与聚类配置默认值切换为 GPU 与 FP16 精度，并严格遵循调用方传入的设备与精度。
  - 重构 DINO 注意力机制为 Pre-scale 架构，将缩放提前至点积前以杜绝 FP16 溢出并减少 16 倍乘法。
  - 放开 LayerNorm FP16 原生调度，消融验证表明可达到 0.999977 余弦相似度并缩减 Engine 体积。
  - 保留 SoftMax 在 FP16 模式下的 FP32 约束，杜绝指数幂极端值溢出。
  - TensorRT 构建期启用 kPREFER_PRECISION_CONSTRAINTS 并更新特征检索单元测试断言。

+++ tests/features/TestImageCluster.cpp
@@ -35,7 +35,7 @@ TEST(ImageClusterTest, DefaultConstructsWithImageFeatureDefaults)
     EXPECT_EQ(cluster.config().model_name, irt::features::ImageCluster::kDefaultModelName);
     EXPECT_EQ(cluster.config().feature_name, irt::features::ImageCluster::kDefaultFeatureName);
     EXPECT_EQ(cluster.config().model_runtime, irt::model::ModelRuntime{});
-    EXPECT_EQ(cluster.config().model_precision, irt::model::ModelPrecision::FP32);
+    EXPECT_EQ(cluster.config().model_precision, irt::model::ModelPrecision::FP16);
     EXPECT_FALSE(cluster.config().use_pca);
     EXPECT_EQ(cluster.config().pca_dim, 0);
     EXPECT_EQ(cluster.config().hdbscan.min_cluster_size, 5);
